backend/processing/requirements.txt
psycopg[binary]>=3.1.18
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=8.0.0
